+++ src/Electron/src/Main/ARCtrlExtensions/ArcFileCreateContracts.fs
@@ -12,4 +12,6 @@ module ArcFileCreateContracts =
         | ArcFiles.Study(study, _) -> study.ToCreateContract(withFolder)
         | ArcFiles.Workflow workflow -> workflow.ToCreateContract(withFolder)
         | ArcFiles.Run run -> run.ToCreateContract(withFolder)
+        | ArcFiles.DataMap(Some parentInfo, dataMap) ->
+            DataMapContracts.createForParent parentInfo dataMap |> Array.singleton
         | unsupportedArcFile -> failwithf "Cannot create ARC file contracts for %A." unsupportedArcFile